pyproject.toml
[project]
name = "lieutenant"
version = "0.1.0"
description = "Encode numbers using the mnemonic major system"
readme = "README.md"
authors = [
    { name = "Emily Prieto", email = "61399347+stygian-ash@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "automata-lib>=9.1.2",
    "diskcache>=5.6.3",
    "platformdirs>=4.5.0",
    "typer>=0.20.0",
]
license = "MIT"
license-files = ["LICENSE"]

[project.scripts]
lieutenant = "lieutenant:main"

[project.urls]
Homepage = "https://github.com/stygian-ash/lieutenant"
Repository = "https://github.com/stygian-ash/lieutenant.git"
Issues = "https://github.com/stygian-ash/lieutenant/issues"

[build-system]
requires = ["uv_build>=0.9.12,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mosestokenizer>=1.2.1",
    "nltk>=3.9.2",
    "pytest>=9.0.1",
]
